Backend/requirements.txt
Django==3.2.6
django-environ==0.8.1
djangorestframework==3.13
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.11.0
black==22.3.0
isort==5.10.1
